==> 3-1_Preprocess - missing values of train data.ipynb <==
# Preprocess of house prices dataset
In the following notebooks we're going to preprocess the data, that is remove missing variables, transform the variables and treat outliers. We're also going to build a specialized pipeline for those transformations.

In this notebook specifically, missing values will be treated.

# import dataset and libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from scipy.stats import pearsonr
%matplotlib inline
pd.plotting.register_matplotlib_converters()
plt.rc('figure', figsize=(16, 6))

orig_data = pd.read_csv("data/train.csv", index_col="Id")

# copying the dataset for analysis
house_data = orig_data.copy()
house_data.head()

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

As seen from univariate and bivariate analysis many of these variables have missing values because they don't have that feature. E.g. it's not possible to determine pool quality for houses which don't have a pool. Although we have to be careful because not all of the missing values have such simple reason.

First let's check if missing values from "Garage" variable are because of such reason. With help there comes a variable named "GarageArea" in which we can check for how many rows it has value of 0 (no garage).

house_data["GarageArea"][house_data["GarageArea"] == 0].count()

It seems to match perfectly with the number of missing values. Let's check further if the rows that have value of 0 for "GarageArea", also have missing values for the rest:

garage_features = ["GarageType", "GarageYrBlt", "GarageFinish", "GarageQual", "GarageCond"]
for feature in garage_features:    
    print(house_data[feature][house_data["GarageArea"] == 0].isnull().sum())

They do! Therefore we can replace them with a category "NoGarage" for now until we decide how to use this variable in the model

for feature in garage_features:
    house_data[feature] = house_data[feature].fillna("NoGarage")

We can now check that none of the garage features have missing variables:

missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

Similarly we can replace values from FireplaceQu, PoolQC, MiscFeature using the same method

house_data["Fireplaces"][house_data['Fireplaces'] == 0].count()

print(house_data['FireplaceQu'][house_data["Fireplaces"] == 0].isnull().sum())

house_data["FireplaceQu"] = house_data["FireplaceQu"].fillna("NoFireplace")

house_data["PoolArea"][house_data['PoolArea'] == 0].count()

print(house_data['PoolQC'][house_data["PoolArea"] == 0].isnull().sum())

house_data["PoolQC"] = house_data["PoolQC"].fillna("NoPool")

house_data["MiscVal"][house_data['MiscVal'] == 0].count()

print(house_data['MiscFeature'][house_data["MiscVal"] == 0].isnull().sum())

# As the number of missing values with MiscVal = 0 is the same as the number of missing values overall 
# we can treat those missing values as having no miscellaneous feature
# There seems to be also two miscellaneous features with no value
house_data["MiscFeature"] = house_data["MiscFeature"].fillna("NoFeature")

As explained in data_description.txt missing values in Alley and Fence mean there is no alley access/ no fence. Therefore we can just replace those values with separate categories.

house_data["Fence"] = house_data["Fence"].fillna("NoFence")
house_data["Alley"] = house_data["Alley"].fillna("NoAlley")

missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

There is a lot of missing values in LotFrontage (over 17%) with no obvious way to replace it. Because of thoroughness in gathering data in other variables it seems unlikely that missing values here are because of random errors. Maybe it was difficult to establish exactly what is a LotFrontage for many houses.

To eliminate missing variables in this case I will impute the value with a general central value. To not introduce data leakage I will do this only after data is split into training and validation.

For MasVnrType I will assume that data was missing because of random error. It will be replaced by the mode, that is lack of Masonry.

For the Electrical I will assume the same, it will also be replaced by the mode.

# checking if both MasVnrType and MasVnrArea have missing values in the same rows
house_data["MasVnrArea"][house_data["MasVnrType"].isnull()].isnull().sum()

house_data["MasVnrArea"] = house_data["MasVnrArea"].fillna(0)
house_data["MasVnrType"] = house_data["MasVnrType"].fillna("None")

house_data["Electrical"] = house_data["Electrical"].fillna(house_data["Electrical"].mode()[0])

missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

We are left only with the basement variables, To fix BsmtFinType1 and BsmtFinType2 we can check their variables which determine their square feet. But we need to be careful as values of 0 square feet will correspond to unfinished basements as well as "No basement". To check how many correspond to something else than Unfinished we can use the following:

house_data["BsmtFinSF1"][house_data['BsmtFinSF1'] == 0].count() - house_data["BsmtFinType1"][house_data['BsmtFinType1'] == "Unf"][house_data['BsmtFinSF1'] == 0].count()

It seems that all of the BsmtFinType1 which aren't unfinished and have 0 square feet area correspond to "No Basement" category

Let's do further checks on BsmtQual, BsmtCond and BsmtFinType1:

house_data["BsmtQual"][house_data['BsmtFinSF1'] == 0].isnull().sum()

house_data["BsmtCond"][house_data['BsmtFinSF1'] == 0].isnull().sum()

house_data["BsmtFinType1"][house_data['BsmtFinSF1'] == 0].isnull().sum()

Therefore we can fill these values with a separate category "NoBsmt"

house_data["BsmtQual"] = house_data["BsmtQual"].fillna("NoBsmt")
house_data["BsmtCond"] = house_data["BsmtCond"].fillna("NoBsmt")
house_data["BsmtFinType1"] = house_data["BsmtFinType1"].fillna("NoBsmt")

missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

house_data["BsmtFinSF2"][house_data['BsmtFinSF2'] == 0].count() - house_data["BsmtFinType2"][house_data['BsmtFinType2'] == "Unf"][house_data['BsmtFinSF2'] == 0].count()

It seems that only 37 out of 38 values can be explained by lack of basement. One of them might a true missing value.

Let's check if missing values in Exposure and FinType2 are related:

house_data["BsmtExposure"][house_data["BsmtFinType2"].isnull()].isnull().sum()

We conclude that for 37 of them every single Basement variable has a missing value. For one case there is a missing value in Exposure column, for the other in BsmtFinType2 column:

house_data["BsmtExposure"][house_data["BsmtFinType2"].notnull()].isnull().sum()

house_data["BsmtExposure"][house_data["BsmtFinType2"].isnull()].notnull().sum()

I will replace those single values by modes:

mode_exposure = house_data["BsmtExposure"].mode()[0]
print(mode_exposure)
mode_type2 = house_data["BsmtFinType2"].mode()[0]
print(mode_type2)

# This throws a warning but indeed it fixes a value in the dataframe. 
# This can be avoided by knowing the exact position of the row, which can be obtained normally.
# This also can be avoided by using df.loc which I will try to do in next cell.
house_data["BsmtExposure"][house_data["BsmtFinType2"].notnull()] = house_data["BsmtExposure"][house_data["BsmtFinType2"].notnull()].fillna(mode_exposure)

# This does not throw any errors and is also faster than chained indexing
mask = house_data["BsmtExposure"].notnull()
house_data.loc[mask, "BsmtFinType2"] = house_data["BsmtFinType2"][mask].fillna(mode_type2)

house_data["BsmtExposure"][house_data["BsmtFinType2"].isnull()].notnull().sum()

Now that we eliminated those actual missing values we can replace the values in BsmtExposure and BsmtFinType2 by category "NoBsmt"

house_data["BsmtExposure"] = house_data["BsmtExposure"].fillna("NoBsmt")
house_data["BsmtFinType2"] = house_data["BsmtFinType2"].fillna("NoBsmt")

missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

So the only missing variable is the LotFrontage, which will be dealt with when we divide the data into train and validation.

Finally we save the data into csv file.

house_data.to_csv("data/train_preprocessed.csv")
