==> 3-2_Preprocess - missing values of test data.ipynb <==
# Preprocess of house prices dataset
In the following notebooks we're going to preprocess the data, that is remove missing variables, transform the variables and treat outliers. We're also going to build a specialized pipeline for those transformations.

In this notebook specifically, we're going to fill missing values in test data.

# import dataset and libraries
import numpy as np
import pandas as pd

orig_data = pd.read_csv("data/test.csv", index_col="Id")

# copying the dataset for analysis
house_data = orig_data.copy()
house_data.head()

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

print(house_data['MiscFeature'][house_data["MiscVal"] == 0].isnull().sum())
print(house_data['PoolQC'][house_data["PoolArea"] == 0].isnull().sum())
print(house_data['FireplaceQu'][house_data["Fireplaces"] == 0].isnull().sum())

It seems there is an additional value in MiscFeature that needs attention, that doesn't have 0 miscellaneous value. Let's print it:

house_data[house_data["MiscVal"] != 0][house_data['MiscFeature'].isnull()]

It appears to be mistakenly assigned to lack of feature since its value is above 0. I will change the value for this feature to "Othr":

house_data.loc[2550, "MiscFeature"] = "Othr"

# As explained in missing values of train data, we can fill those features immediately:
house_data["PoolQC"] = house_data["PoolQC"].fillna("NoPool")
house_data["MiscFeature"] = house_data["MiscFeature"].fillna("NoFeature")
house_data["Fence"] = house_data["Fence"].fillna("NoFence")
house_data["Alley"] = house_data["Alley"].fillna("NoAlley")
house_data["FireplaceQu"] = house_data["FireplaceQu"].fillna("NoFireplace")

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

From the rest we can easily set most of the missing values to modes:

mode_replacement = ["MSZoning", "Utilities", "Exterior1st", "Exterior2nd", "BsmtFullBath", "BsmtHalfBath", 
                    "KitchenQual", "Functional", "SaleType"]

for feature in mode_replacement:
    house_data[feature] = house_data[feature].fillna(house_data[feature].mode()[0])

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

Let's first fill the missing values in "Mas" features:

house_data["MasVnrArea"][house_data["MasVnrType"].isnull()].isnull().sum()

house_data[house_data["MasVnrType"].isnull()][house_data["MasVnrArea"].notnull()][["MasVnrArea", "MasVnrType"]]

Since the area is known I cannot assign type to None type. I will instead fill it with the mode out of non-None values

house_data.loc[2611, "MasVnrType"] = house_data["MasVnrType"][house_data["MasVnrType"] != "None"].mode()[0]

# the rest will be filled with 0s
house_data["MasVnrArea"] = house_data["MasVnrArea"].fillna(0)
house_data["MasVnrType"] = house_data["MasVnrType"].fillna("None")

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

for feature in ["GarageYrBlt", "GarageFinish", "GarageQual", "GarageCond"]:
    print(house_data[feature][house_data["GarageType"].isnull()].isnull().sum())

Therefore we can immediately fill those 76 rows with "NoGarage" value

mask = house_data["GarageType"].isnull()

for feature in ["GarageYrBlt", "GarageFinish", "GarageQual", "GarageCond"]:
    house_data.loc[mask, feature] = house_data[feature][mask].fillna("NoGarage")

house_data["GarageType"] = house_data["GarageType"].fillna("NoGarage")

Now, the rest of missing values from Garage

print(house_data["GarageYrBlt"][house_data["GarageFinish"].isnull()].isnull().sum())
print(house_data["GarageYrBlt"][house_data["GarageQual"].isnull()].isnull().sum())
print(house_data["GarageYrBlt"][house_data["GarageCond"].isnull()].isnull().sum())

It seems that those 2 missing values correspond to the same row, let's print it:

garage_features = ["GarageType", "GarageYrBlt", "GarageFinish", "GarageQual", "GarageCond", "GarageCars", "GarageArea"]
house_data[house_data["GarageFinish"].isnull()][garage_features]

Since there is no obvious way of filling them in, I'm going to put mode in each of the features.

In case of GarageArea I'm going to replace the value by its geometrical mean (it's almost geometric mean since I'm using log(1+x) transformation and not log(x)) since it comes into the model by means of a logarithm.

for feature in ["GarageYrBlt", "GarageFinish", "GarageQual", "GarageCond", "GarageCars"]:
    house_data[feature] = house_data[feature].fillna(house_data[feature].mode()[0])

geometric_mean = np.exp(np.mean(np.log(house_data["GarageArea"] + 1))) - 1
house_data["GarageArea"] = house_data["GarageArea"].fillna(geometric_mean)

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

Now we're only left with Basement features since we don't use LotFrontage in our models. From the outlook it seems that one row doesn't contain information about area and about 42 rows have categorical features missing. Let's check if this is the case: 

bsmt_features = ["BsmtQual", "BsmtCond", "BsmtExposure", "BsmtFinType1", "BsmtFinSF1", "BsmtFinType2", "BsmtFinSF2", 
                 "BsmtUnfSF", "TotalBsmtSF"]
mask = house_data["TotalBsmtSF"].isnull()
house_data[mask][bsmt_features]

So for this one row, no information about basement is available. I will therefore assume that it didn't have a basement at all.

for feature in ["BsmtQual", "BsmtCond", "BsmtExposure", "BsmtFinType1", "BsmtFinType2"]:
    house_data.loc[2121, feature] = "NoBsmt"
for feature in ["BsmtFinSF1", "BsmtFinSF2", "BsmtUnfSF", "TotalBsmtSF"]:
    house_data.loc[2121, feature] = 0

mask = house_data["BsmtFinType1"].isnull()
house_data[mask][bsmt_features]

Above means that all 41 of those houses, didn't have a basement.

# filling those 41 rows
mask = house_data["BsmtFinType1"].isnull()

for feature in ["BsmtQual", "BsmtCond", "BsmtExposure", "BsmtFinType2"]:
    house_data.loc[mask, feature] = house_data[mask][feature].fillna("NoBsmt")

house_data["BsmtFinType1"] = house_data["BsmtFinType1"].fillna("NoBsmt")

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

mask = house_data["BsmtCond"].isnull()

print(house_data[mask][["BsmtQual", "BsmtCond", "BsmtExposure"]])

mask = house_data["BsmtQual"].isnull()

print(house_data[mask][["BsmtQual", "BsmtCond", "BsmtExposure"]])

mask = house_data["BsmtExposure"].isnull()

print(house_data[mask][["BsmtQual", "BsmtCond", "BsmtExposure"]])

We conclude they are single instances of missing variables. Therefore I will fill them with modes.

for feature in ["BsmtQual", "BsmtCond", "BsmtExposure"]:
    house_data[feature] = house_data[feature].fillna(house_data[feature].mode()[0])

# checking missing values in each column
missing_val_count_by_column = house_data.isnull().sum()
print(missing_val_count_by_column[missing_val_count_by_column > 0])

# So we have a clear dataset. Let's save it to a file:
house_data.to_csv("data/test_preprocessed.csv")
